feat: database 관련 의존성 주입
