+++ src/main/java/codesquad/springcafe/repository/JdbcArticleRepository.java
@@ -1,0 +1,2 @@
+package codesquad.springcafe.repository;public class JdbcArticleRepository {
+}
