+++ src/main/java/codesquad/springcafe/repository/JdbcUserRepository.java
@@ -1,0 +1,4 @@
+package codesquad.springcafe.repository;
+
+public class JdbcUserRepository {
+}
